added SparkVersionUtils
